backend/requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-multipart==0.0.9
python-dotenv==1.0.1

# ── HTTP / HTML ────────────────────────────────────────────────────────────
httpx==0.27.0
requests==2.31.0
beautifulsoup4==4.12.3

# ── URL feature extraction ─────────────────────────────────────────────────
tldextract==5.1.2
python-whois==0.9.4
levenshtein==0.25.1

# ── ML core ───────────────────────────────────────────────────────────────
scikit-learn==1.4.2
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
imbalanced-learn==0.12.3

# ── Deep learning ─────────────────────────────────────────────────────────
torch==2.3.0
torchvision==0.18.0
transformers==4.40.2
datasets==2.19.1
accelerate==0.30.0

# ── Screenshots ───────────────────────────────────────────────────────────
playwright==1.44.0
Pillow==10.3.0

# ── Database ──────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
aiosqlite==0.20.0

# ── XAI ───────────────────────────────────────────────────────────────────
shap==0.45.1
lime==0.2.0.1
transformers-interpret

# ── Robustness attacks ─────────────────────────────────────────────────────
# (no extra deps — uses numpy, PIL, re, urllib already above)

# ── Evaluation & visualisation ─────────────────────────────────────────────
matplotlib==3.9.0
seaborn==0.13.2
scipy==1.13.0

# ── Profiling ─────────────────────────────────────────────────────────────
psutil==5.9.8
